DRF - Django REST Framework/Proj_Cursos_API/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
jsonpath==0.82
Markdown==3.4.3
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
urllib3==2.0.2
